requirements.txt
streamlit
pandas
numpy
azure-storage-blob
